update productBox in shop
