+++ core/front-end/account/index.html
@@ -5,9 +5,6 @@
     <meta http-equiv="X-UA-Compatible" content="IE=Edge"/>
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
 
-    <!-- style -->
-    <link rel="stylesheet" href="/src/css/global.css">
-
     <!-- title -->
     <title></title>
