+++ core/front-end/src/js/pages/shop.js
@@ -1,5 +1,5 @@
 import {getAllProducts} from "@/js/api/auth.js";
-import {createProductBox} from "@/js/main/main.js"
+import productBox from "@/js/component/product-box/productBox.js";
 
 export function init() {
     const productsWrapper = document.getElementById('products-wrapper');
@@ -11,12 +11,24 @@ export function init() {
 
             if (products[0]) {
                 const fragment = document.createDocumentFragment();
+                const res = await getAllProducts()
+
+                const allProductsTemplate = res.map(product => {
+                    console.log(productBox(product));
+                    fragment.appendChild(productBox(product));
+                });
+
+                console.log(allProductsTemplate);
+
+                console.log(fragment);
+
                 productsWrapper.appendChild(fragment);
             } else {
                 textMessage.innerHTML = "<h2 class='w-full text-center'>هیچ محصولی وجود ندارد</h2>"
             }
         } catch (e) {
             if (e instanceof Error) {
+                console.log(e)
                 textMessage.innerHTML = "<h2 class='w-full text-center'>در حال حاضر سرور در دسترس نمیباشد</h2>"
             } else {
                 console.error("more: ", e)
